+++ R/facet-wrap.r
@@ -14,7 +14,7 @@ facet_rep_wrap <- function(..., repeat.tick.labels=FALSE) {
 }
 
 
-#' @rdname splot-ggproto
+#' @rdname lemon-ggproto
 #' @keywords internal
 #' @format NULL
 #' @usage NULL
@@ -48,9 +48,9 @@ FacetWrapRepeatLabels <- ggplot2::ggproto('FacetWrapRepeatLabels',
     panel_order <- order(layout$ROW, layout$COL)
     layout <- layout[panel_order, ]
     panels <- panels[panel_order]
-    panel_pos <- ggplot2:::convertInd(layout$ROW, layout$COL, nrow)
+    panel_pos <- convertInd(layout$ROW, layout$COL, nrow)
 
-    axes <- ggplot2:::render_axes(ranges, ranges, coord, theme, transpose = TRUE)
+    axes <- render_axes(ranges, ranges, coord, theme, transpose = TRUE)
 
     labels_df <- layout[names(params$facets)]
     attr(labels_df, "facet") <- "wrap"
@@ -76,7 +76,7 @@ FacetWrapRepeatLabels <- ggplot2::ggproto('FacetWrapRepeatLabels',
     empty_table <- matrix(list(ggplot2::zeroGrob()), nrow = nrow, ncol = ncol)
     panel_table <- empty_table
     panel_table[panel_pos] <- panels
-    empties <- apply(panel_table, c(1,2), function(x) ggplot2:::is.zero(x[[1]]))
+    empties <- apply(panel_table, c(1,2), function(x) is.zero(x[[1]]))
     panel_table <- gtable_matrix("layout", panel_table,
      widths = unit(rep(1, ncol), "null"),
      heights = unit(rep(aspect_ratio, nrow), "null"), respect = respect, clip = "on", z = matrix(1, ncol = ncol, nrow = nrow))
@@ -119,10 +119,10 @@ FacetWrapRepeatLabels <- ggplot2::ggproto('FacetWrapRepeatLabels',
       first_row <- which(apply(empties, 1, any))[1] - 1
       first_col <- which(apply(empties, 2, any))[1] - 1
       row_panels <- which(layout$ROW == first_row & layout$COL > first_col)
-      row_pos <- ggplot2:::convertInd(layout$ROW[row_panels], layout$COL[row_panels], nrow)
+      row_pos <- convertInd(layout$ROW[row_panels], layout$COL[row_panels], nrow)
       row_axes <- axes$x$bottom[layout$SCALE_X[row_panels]]
       col_panels <- which(layout$ROW > first_row & layout$COL == first_col)
-      col_pos <- ggplot2:::convertInd(layout$ROW[col_panels], layout$COL[col_panels], nrow)
+      col_pos <- convertInd(layout$ROW[col_panels], layout$COL[col_panels], nrow)
       col_axes <- axes$y$right[layout$SCALE_Y[col_panels]]
       if (params$strip.position == "bottom" &&
           theme$strip.placement != "inside" &&
@@ -139,10 +139,10 @@ FacetWrapRepeatLabels <- ggplot2::ggproto('FacetWrapRepeatLabels',
         axis_mat_y_right[col_pos] <- col_axes
       }
     }
-    panel_table <- ggplot2:::weave_tables_row(panel_table, axis_mat_x_top, -1, axis_height_top, "axis-t", 3)
-    panel_table <- ggplot2:::weave_tables_row(panel_table, axis_mat_x_bottom, 0, axis_height_bottom, "axis-b", 3)
-    panel_table <- ggplot2:::weave_tables_col(panel_table, axis_mat_y_left, -1, axis_width_left, "axis-l", 3)
-    panel_table <- ggplot2:::weave_tables_col(panel_table, axis_mat_y_right, 0, axis_width_right, "axis-r", 3)
+    panel_table <- weave_tables_row(panel_table, axis_mat_x_top, -1, axis_height_top, "axis-t", 3)
+    panel_table <- weave_tables_row(panel_table, axis_mat_x_bottom, 0, axis_height_bottom, "axis-b", 3)
+    panel_table <- weave_tables_col(panel_table, axis_mat_y_left, -1, axis_width_left, "axis-l", 3)
+    panel_table <- weave_tables_col(panel_table, axis_mat_y_right, 0, axis_width_right, "axis-r", 3)
 
     strip_padding <- convertUnit(theme$strip.switch.pad.wrap, "cm")
     strip_name <- paste0("strip-", substr(params$strip.position, 1, 1))
@@ -158,10 +158,10 @@ FacetWrapRepeatLabels <- ggplot2::ggproto('FacetWrapRepeatLabels',
         strip_pad <- axis_height_bottom
       }
       strip_height <- unit(apply(strip_mat, 1, max_height), "cm")
-      panel_table <- ggplot2:::weave_tables_row(panel_table, strip_mat, placement, strip_height, strip_name, 2, "on")
+      panel_table <- weave_tables_row(panel_table, strip_mat, placement, strip_height, strip_name, 2, "on")
       if (!inside) {
         strip_pad[unclass(strip_pad) != 0] <- strip_padding
-        panel_table <- ggplot2:::weave_tables_row(panel_table, row_shift = placement, row_height = strip_pad)
+        panel_table <- weave_tables_row(panel_table, row_shift = placement, row_height = strip_pad)
       }
     } else {
       inside <- (theme$strip.placement.y %||% theme$strip.placement %||% "inside") == "inside"
@@ -174,12 +174,82 @@ FacetWrapRepeatLabels <- ggplot2::ggproto('FacetWrapRepeatLabels',
       }
       strip_pad[unclass(strip_pad) != 0] <- strip_padding
       strip_width <- unit(apply(strip_mat, 2, max_width), "cm")
-      panel_table <- ggplot2:::weave_tables_col(panel_table, strip_mat, placement, strip_width, strip_name, 2, "on")
+      panel_table <- weave_tables_col(panel_table, strip_mat, placement, strip_width, strip_name, 2, "on")
       if (!inside) {
         strip_pad[unclass(strip_pad) != 0] <- strip_padding
-        panel_table <- ggplot2:::weave_tables_col(panel_table, col_shift = placement, col_width = strip_pad)
+        panel_table <- weave_tables_col(panel_table, col_shift = placement, col_width = strip_pad)
       }
     }
     panel_table
   }
 )
+
+# Helpers -----------------------------------------------------------------
+
+sanitise_dim <- function(n) {
+  xname <- paste0("`", deparse(substitute(n)), "`")
+  if (length(n) == 0) {
+    if (!is.null(n)) {
+      warning(xname, " has length zero and will be treated as NULL.",
+              call. = FALSE)
+    }
+    return(NULL)
+  }
+  if (length(n) > 1) {
+    warning("Only the first value of ", xname, " will be used.", call. = FALSE)
+    n <- n[1]
+  }
+  if (!is.numeric(n) || (!is.na(n) && n != round(n))) {
+    warning("Coercing ", xname, " to be an integer.", call. = FALSE)
+    n <- as.integer(n)
+  }
+  if (is.na(n) || n < 1) {
+    warning(xname, " is missing or less than 1 and will be treated as NULL.",
+            call. = FALSE)
+    return(NULL)
+  }
+  n
+}
+
+wrap_dims <- function(n, nrow = NULL, ncol = NULL) {
+  if (is.null(ncol) && is.null(nrow)) {
+    rc <- grDevices::n2mfrow(n)
+    nrow <- rc[2]
+    ncol <- rc[1]
+  } else if (is.null(ncol)) {
+    ncol <- ceiling(n / nrow)
+  } else if (is.null(nrow)) {
+    nrow <- ceiling(n / ncol)
+  }
+  stopifnot(nrow * ncol >= n)
+
+  c(nrow, ncol)
+}
+convertInd <- function(row, col, nrow) {
+  (col - 1) * nrow + row
+}
+
+weave_tables_col <- function(table, table2, col_shift, col_width, name, z = 1, clip = "off") {
+  panel_col <- panel_cols(table)$l
+  panel_row <- panel_rows(table)$t
+  for (i in rev(seq_along(panel_col))) {
+    col_ind <- panel_col[i] + col_shift
+    table <- gtable::gtable_add_cols(table, col_width[i], pos = col_ind)
+    if (!missing(table2)) {
+      table <- gtable::gtable_add_grob(table, table2[, i], t = panel_row, l = col_ind + 1, clip = clip, name = paste0(name, "-", seq_along(panel_row), "-", i), z = z)
+    }
+  }
+  table
+}
+weave_tables_row <- function(table, table2, row_shift, row_height, name, z = 1, clip = "off") {
+  panel_col <- panel_cols(table)$l
+  panel_row <- panel_rows(table)$t
+  for (i in rev(seq_along(panel_row))) {
+    row_ind <- panel_row[i] + row_shift
+    table <- gtable::gtable_add_rows(table, row_height[i], pos = row_ind)
+    if (!missing(table2)) {
+      table <- gtable::gtable_add_grob(table, table2[i, ], t = row_ind + 1, l = panel_col, clip = clip, name = paste0(name, "-", seq_along(panel_col), "-", i), z = z)
+    }
+  }
+  table
+}