+++ R/legends.r
@@ -10,7 +10,7 @@ NULL
 #' to the ggplot2 object, before calling \code{g_legend}.
 #'
 #' An alternative method for extracting the legend is using
-#' \code{gridExtra::\link[gtable]{gtable_filter}}:
+#' \code{gtable::\link[gtable]{gtable_filter}}:
 #'
 #' \preformatted{
 #'   gtable_filter(ggplotGrob(a.ggplot.obj), 'guide-box')
@@ -24,13 +24,14 @@ NULL
 #' @param a.gplot ggplot2 or gtable object.
 #' @return gtable (grob) object. Draw with \code{\link[grid]{grid.draw}}.
 #' @export
-#' @author Stack Overflow
+#' @author \href{http://baptiste.github.io/}{Baptiste Auguié}
 #' @import ggplot2
 #' @import gtable
 #' @seealso \code{\link{grid_arrange_shared_legend}}, \code{\link{reposition_legend}},
 #'          \code{\link[gtable]{gtable_filter}}
 #' @examples
 #' library(ggplot2)
+#' library(gtable)
 #' library(grid)
 #' library(gridExtra)
 #' library(gtable)
@@ -71,8 +72,9 @@ g_legend<-function(a.gplot){
 #' Share a legend between multiple plots
 #'
 #' Extract legend, combines plots using \code{\link[gridExtra]{arrangeGrob}},
-#' and places legend underneath.
+#' and places legend in a margin.
 #'
+#' 
 #'
 #' @param ... ggplot2 objects. Their legends are automatically hidden.
 #'            The legend is taken from the first argument.
@@ -81,12 +83,14 @@ g_legend<-function(a.gplot){
 #' @param position 'bottom' or 'right' for positioning legend.
 #' @param plot Logical, when \code{TRUE} (default), draws combined plot on a
 #'             new page.
-#' @return gtable of combined plot, invisibly.
-#' @source
+#' @return gtable of combined plot, invisibly. 
+#'   Draw  gtable object using \code{\link[grid]{grid.draw}}.
+#' @author
 #'   Originally brought to you by \href{http://rpubs.com/sjackman}{Shaun Jackman}
 #'   (\href{http://rpubs.com/sjackman/grid_arrange_shared_legend}{original}),
-#'   and further improved by \code{baptiste} at
-#'   \url{https://github.com/tidyverse/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs}
+#'   and further improved by \href{http://baptiste.github.io/}{Baptiste Auguié}  at
+#'   \url{https://github.com/tidyverse/ggplot2/wiki/Share-a-legend-between-two-ggplot2-graphs}.
+#'   Stefan McKinnon Edwards added left and top margins.
 #' @import ggplot2 gridExtra grid
 #' @export
 #' @seealso \code{\link{g_legend}}, \code{\link{reposition_legend}}
@@ -146,7 +150,7 @@ grid_arrange_shared_legend <- function(...,
   # return gtable invisibly
   invisible(combined)
 }
-
+arrangeGrob <- gridExtra::arrangeGrob
 
 
 #' Reposition a legend onto a panel
@@ -158,13 +162,17 @@ grid_arrange_shared_legend <- function(...,
 #' To modify the look of the legend, use themes and the natural ggplot functions
 #' found in \code{\link[ggplot2]{guide_legend}}.
 #'
-#' Panel name is \code{panel}, but when using facets, it typically takes the
-#' form \code{panel-{col}-{row}}. Use
-#' \preformatted{
-#' ggplot_gtable(ggplot_build(aplot))
-#' }
-#' to build a \code{\link[gtable]{gtable}} object, and print it to look at the
-#' names.
+#' Panel name is by default \code{panel}, but when using facets it typically 
+#' takes the form \code{panel-{col}-{row}}, but not for wrapped facets.
+#' Either print result from \code{\link[ggplot2]{ggplotGrob}} or use
+#' \code{\link{gtable_show_names}} to display all the names of the gtable 
+#' object.
+#'
+#' \code{panel} takes multiple names, and will then use these components'
+#' extremes for placing the legend.
+#' 
+#' If \code{panel} is an integer vector of length 2 or 4, these elements are
+#' used directly for top-left and bottom-right coordinates.
 #'
 #' @param aplot a ggplot2 or gtable object.
 #' @param position Where to place the legend in the panel.
@@ -172,11 +180,13 @@ grid_arrange_shared_legend <- function(...,
 #' @param legend The legend to place, if \code{NULL} (default),
 #'               it is extracted from \code{aplot} if
 #'               this is a ggplot2 object.
-#' @param panel Name of panel in gtable.
+#' @param panel Name of panel in gtable. See description.
 #' @param x horisontal coordinate of legend, with 0 at left.
 #' @param y vertical coordiante of legend, with 0 at bottom.
 #' @param just 'Anchor point' of legend; it is this point of the legend that is
 #'             placed at the \code{x} and \code{y} coordinates.
+#' @param name,clip,z Parameters forwarded to 
+#'             \code{\link[gtable]{gtable_add_grob}}.
 #' @param plot Logical, when \code{TRUE} (default), draws plot with legend
 #'             repositioned on a new page.
 #' @return
@@ -209,6 +219,9 @@ reposition_legend <- function(aplot,
                              x=NULL,
                              y=NULL,
                              just=NULL,
+                             name='guide-box',
+                             clip='on',
+                             z=Inf,
                              plot=TRUE) {
 
   # Work out positioning
@@ -242,23 +255,47 @@ reposition_legend <- function(aplot,
   if (!inherits(aplot, 'gtable'))
     aplot <- ggplot_gtable(ggplot_build(aplot + theme(legend.position='hidden')))
 
-  legend$vp <- viewport(x=x, y=y, just=just,  width=sum(legend$widths), height=sum(legend$heights))
-
-  pn <- which(aplot$layout$name == panel)
-  if (length(pn) == 0) stop('Could not find panel named `',panel,'`.')
-  if (length(pn) > 1 ) {
-    warning('Multiple panels named `', panel, '`; using first found.')
-    pn <- pn[1]
+  # Update name if already found. No idea why this is necessary.
+  if (any(grepl(name, aplot$layout$name))) {
+    add <- gregexpr(paste0('(?!', name, '-?)([[:digit:]]+)'), aplot$layout$name, perl = TRUE)
+    add <- regmatches(aplot$layout$name, add)
+    add <- as.integer(unlist(add))
+    if (length(add) > 0) {
+      add <- max(add) + 1
+    } else {
+      add <- 1
+    }
+    name <- paste(name, add, sep='-')
   }
-  #if (inherits(aplot$grobs[[pn]], 'zeroGrob')) {
-  #  aplot$grobs[[pn]] <- legend
-  #} else {
-  #  aplot$grobs[[pn]]$vp <- viewport(x=unit(0, 'npc'), y=unit(0, 'npc'), just=c(0,0), name=panel)
-  #  aplot$grobs[[pn]]$children <- gList(aplot$grobs[[pn]]$children, `guide-box`=legend)
-  #  aplot$grobs[[pn]]$childrenOrder <- append(aplot$grobs[[pn]]$childrenOrder, 'guide-box')
-  #}
-  aplot <- with(aplot$layout[pn,], gtable_add_grob(aplot, legend, t=t, l=l, r=r, b=b))
-
+  
+  legend$vp <- viewport(x=x, y=y, just=just,  width=sum(legend$widths), height=sum(legend$heights))
+  legend$name <- name
+  
+  if (is.character(panel)) {
+    pn <- which(aplot$layout$name %in% panel)
+    if (length(pn) == 0) stop('Could not find panel named `',panel,'`.')
+  
+    aplot <- with(aplot$layout[pn,], 
+                  gtable_add_grob(aplot, 
+                                  legend, 
+                                  t = min(t), 
+                                  l = min(l), 
+                                  r = max(r), 
+                                  b = max(b),
+                                  name = legend$name
+                                  ))
+  } else if ((is.numeric(panel) | is.integer(panel)) & length(panel) %in% c(2,4)) {
+    panel <- rep(as.integer(panel), length.out=4)
+    aplot <- gtable_add_grob(aplot, 
+                             legend, 
+                             t = panel[1], 
+                             l = panel[2], 
+                             b = panel[3], 
+                             r = panel[4], 
+                             name = legend$name
+                             )
+  }
+  
   if (plot) {
     grid.newpage()
     grid.draw(aplot)